requirements.txt
# Core dependencies
python-dotenv==1.0.0
pinecone-client==3.1.0
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.13
transformers==4.36.0
torch

# Data processing
pandas
numpy
scikit-learn

# API and UI
streamlit
requests

# Evaluation
rouge-score
bert-score

# Utilities
tqdm
jsonlines
